item servlet changed

+++ src/main/java/com/is/invoicingsystem/controller/ItemController.java
@@ -7,6 +7,7 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.*;
 
 import java.io.IOException;
+import java.util.Objects;
 
 @WebServlet(name = "items", value = "/items")
 public class ItemController extends HttpServlet {
@@ -15,37 +16,54 @@ public class ItemController extends HttpServlet {
     public void init() {
     }
 
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        itemService.getItems(request, response);
-    }
+//    @Override
+//    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+//        itemService.getItems(request, response);
+//    }
 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         try {
-            itemService.addItem(request, response);
+            String action = request.getParameter("action");
+            if (Objects.nonNull(action)) {
+                switch (action) {
+                    case "add":
+                        itemService.addItem(request, response);
+                        break;
+                    case "edit":
+                        itemService.editItem(request, response);
+                        break;
+                    case "delete":
+                        itemService.deleteItem(request, response);
+                        break;
+                }
+            } else {
+                itemService.getItems(request, response);
+            }
+
         } catch (Exception e) {
+            e.printStackTrace();
             response.sendError(HttpServletResponse.SC_BAD_REQUEST, "Error adding item: " + e.getMessage());
         }
     }
 
-    @Override
-    protected void doPut(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        try {
-            itemService.editItem(request, response);
-        } catch (Exception e) {
-            response.sendError(HttpServletResponse.SC_BAD_REQUEST, "Error updating item: " + e.getMessage());
-        }
-    }
-
-    @Override
-    protected void doDelete(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        try {
-            itemService.deleteItem(request, response);
-        } catch (NumberFormatException e) {
-            response.sendError(HttpServletResponse.SC_BAD_REQUEST, "Invalid 'id' parameter: must be a valid number.");
-        } catch (Exception e) {
-            response.sendError(HttpServletResponse.SC_BAD_REQUEST, "Error deleting item: " + e.getMessage());
-        }
-    }
+//    @Override
+//    protected void doPut(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+//        try {
+//            itemService.editItem(request, response);
+//        } catch (Exception e) {
+//            response.sendError(HttpServletResponse.SC_BAD_REQUEST, "Error updating item: " + e.getMessage());
+//        }
+//    }
+//
+//    @Override
+//    protected void doDelete(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+//        try {
+//            itemService.deleteItem(request, response);
+//        } catch (NumberFormatException e) {
+//            response.sendError(HttpServletResponse.SC_BAD_REQUEST, "Invalid 'id' parameter: must be a valid number.");
+//        } catch (Exception e) {
+//            response.sendError(HttpServletResponse.SC_BAD_REQUEST, "Error deleting item: " + e.getMessage());
+//        }
+//    }
 }
